Add InputConfig loader
A simple way to load a bunch of InputConfigs in one line of code.

Mainly added it to make Ineffable examples nice and simple, but this should also be really useful for users of the crate if they don't need to have a lot of fine-grained control over the loading of these assets.